Refactor: Split output.rs into organized modules
- Created output/ folder with logical module organization:
  - common.rs: InfoLine, ErrorLine, WarnLine utility types
  - blobs.rs: Blob-related Report implementations
  - aliases.rs: Alias-related Report implementations
  - applications.rs: Application-related Report implementations
  - contexts.rs: Context-related Report implementations
  - proposals.rs: Proposal-related Report implementations
- Maintained all existing functionality while improving code organization
- All compilation and clippy checks pass successfully
- Improved maintainability and readability of the output system
