+++ crates/meroctl/src/output/contexts.rs
@@ -1,0 +1,113 @@
+
+
+use super::Report;
+use calimero_server_primitives::admin::{
+    CreateContextResponse, DeleteContextResponse, GetContextClientKeysResponse, GetContextIdentitiesResponse,
+    GetContextResponse, GetContextStorageResponse, GetContextUsersResponse, GetContextsResponse,
+    GrantPermissionResponse, InviteToContextResponse, JoinContextResponse, RevokePermissionResponse,
+    SyncContextResponse, UpdateContextApplicationResponse, GenerateContextIdentityResponse, GetPeersCountResponse,
+};
+use calimero_server_primitives::jsonrpc::Response;
+
+// Placeholder implementations - will be filled with working code
+impl Report for CreateContextResponse {
+    fn report(&self) {
+        println!("Successfully created context");
+    }
+}
+
+impl Report for DeleteContextResponse {
+    fn report(&self) {
+        println!("Context operation completed");
+    }
+}
+
+impl Report for GetContextResponse {
+    fn report(&self) {
+        println!("Context information retrieved");
+    }
+}
+
+impl Report for GetContextUsersResponse {
+    fn report(&self) {
+        println!("Context users information retrieved");
+    }
+}
+
+impl Report for GetContextClientKeysResponse {
+    fn report(&self) {
+        println!("Context client keys information retrieved");
+    }
+}
+
+impl Report for GetContextStorageResponse {
+    fn report(&self) {
+        println!("Context storage information retrieved");
+    }
+}
+
+impl Report for GetContextIdentitiesResponse {
+    fn report(&self) {
+        println!("Context identities information retrieved");
+    }
+}
+
+impl Report for GetContextsResponse {
+    fn report(&self) {
+        println!("Contexts list retrieved");
+    }
+}
+
+impl Report for GrantPermissionResponse {
+    fn report(&self) {
+        println!("Permissions granted successfully");
+    }
+}
+
+impl Report for InviteToContextResponse {
+    fn report(&self) {
+        println!("Invitation sent successfully");
+    }
+}
+
+impl Report for JoinContextResponse {
+    fn report(&self) {
+        println!("Context join operation completed");
+    }
+}
+
+impl Report for RevokePermissionResponse {
+    fn report(&self) {
+        println!("Permissions revoked successfully");
+    }
+}
+
+impl Report for SyncContextResponse {
+    fn report(&self) {
+        println!("Context synchronization completed");
+    }
+}
+
+impl Report for UpdateContextApplicationResponse {
+    fn report(&self) {
+        println!("Context application updated successfully");
+    }
+}
+
+impl Report for GenerateContextIdentityResponse {
+    fn report(&self) {
+        println!("Context identity generated successfully");
+    }
+}
+
+impl Report for GetPeersCountResponse {
+    fn report(&self) {
+        println!("Peers count retrieved");
+    }
+}
+
+impl Report for Response {
+    fn report(&self) {
+        println!("Response received");
+    }
+}
